refactoring existentent code for expid where it is used the path library and a new unit test code was generate to make sure everything works as expected
